% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importF.R
\name{asGR}
\alias{asGR}
\title{Creates Granges object from genotyping results}
\usage{
asGR(myDNA)
}
\arguments{
\item{myDNA}{data.frame. Results of read.data for genotyping results}

\item{type}{(character) info about the company used to genotype data:
myHeritage, 23andMe, AncestryDNA, Illumina}
}
\description{
Creates Granges object from genotyping results
}
\details{
outputs granges
}
\author{
Inga Patarcic
}
\keyword{internal}
