% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{blueEyes}
\alias{blueEyes}
\title{Function to screen for production of ACTN3 gene}
\usage{
blueEyes(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for production of ACTN3 gene
}
\details{
SOURCE 1: 23andMe report
INFO:
  The genetic marker in this report is located near a gene called OCA2 that
affects how much brown pigment your cells produce. People with 1 or 2 copies
of the A variant of this marker tend to have more brown pigment in their eyes,
so they are likely to have darker eyes.
AA/AG - likely brown/hazel eyes, GG - likely blue/green eyes
NOTE! MyHeritage genotypes do not report this SNP, however missing genotypes
can be imputed and accessed by DNALand imputation tool

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
 library(myDNA)
 library(plyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
 blueEyes(myDNA)
}
}
\author{
Inga Patarcic
}
