% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{blueEyesExtended}
\alias{blueEyesExtended}
\title{Function to screen for testing eye color}
\usage{
blueEyesExtended(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for testing eye color
}
\details{
Source 2: SNPedia for rs12913832
resuts source 2: rs12913832 is also part of a haplotype spanning
166kB on chromosome 15, defined by 13 SNPs listed below,
that is found in 97% of all Caucasians with blue eyes.
The "h-1" haplotype found in homozygous state in 97% of individuals
with blue eye color is composed as follows [PMID 18172690]:

rs4778241(C)
rs1129038(A)
rs12593929(A)
rs12913832(G)  OCA2/HERC2 has a score of 51.5 and an estimated allelic OR of 8.43
rs7183877(C)
rs3935591(G)
rs7170852(A)
rs2238289(T)
rs3940272(C)
rs8028689(T)
rs2240203(A)
rs11631797(G)
rs916977(G)
ADDED from SNPedia: rs1667394 increases susceptibility to Blond rather than
brown hair 4.94 times for carriers of the A allele

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
 library(myDNA)
 library(plyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
 blueEyesExtended(myDNA)
}
}
\author{
Inga Patarcic
}
