% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{celiacDisease23AndMe}
\alias{celiacDisease23AndMe}
\title{Function to screen for celiac disease based on 23andMe tested SNPs}
\usage{
celiacDisease23AndMe(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for celiac disease based on 23andMe tested SNPs
}
\details{
SCREEN 1: rs2187668 (T) and rs7454108 (G)
SOURCE: https://blog.23andme.com/health-traits/new-23andme-report-celiac-disease/

NOTE! MyHeritage genotypes do not report this SNP, however missing genotypes
can be imputed and accessed by DNALand imputation tool

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{
\dontrun{
# example myHeritage
 library(myDNA)
 library(dplyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
  celiacDisease23AndMe(myDNA)
}
}
\author{
Inga Patarcic
}
