% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{celiacDiseaseExtended}
\alias{celiacDiseaseExtended}
\title{Function to screen for celiac disease based on 6 CD associated SNPs}
\usage{
celiacDiseaseExtended(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for celiac disease based on 6 CD associated SNPs
}
\details{
SOURCE: https://www.mygenefood.com/genetics-celiac-disease-need-know/
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2386975/
  DQ2.5	rs2187668	T (0.09)
DQ8	rs7454108	G (0.18)
DQ2.2	rs2395182	T (0.71)
rs7775228	G (0.10)
rs4713586	G (0.025)
DQ7	rs4639334	A (0.09)
NOTE! MyHeritage genotypes do not report this SNP, however missing genotypes
can be imputed and accessed by DNALand imputation tool

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{
\dontrun{
# example myHeritage
 library(myDNA)
 library(plyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
  celiacDiseaseExtended(myDNA)
}
}
\author{
Inga Patarcic
}
