% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{cofeeConsumption}
\alias{cofeeConsumption}
\title{Function to screen for higher coffee consumption}
\usage{
cofeeConsumption(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for higher coffee consumption
}
\details{
SOURCE: 23andMe report
INFO:
  Report is based on genetic variants near two genes that play a role in how
your body handles caffeine. The first gene, CYP1A2, contains instructions for
an enzyme that breaks down 95% of the caffeine you consume.
The second gene, AHR (rs4410790), contains instructions for a protein that ramps
up production of the CYP1A2 enzyme. Variants in these genes may affect
how quickly the body breaks down and clears away caffeine.
NOTE! MyHeritage genotypes do not report this SNP, however missing genotypes
can be imputed and accessed by DNALand imputation tool
The CYP1A2 gene (rs2472297) contains instructions for an enzyme that breaks
down many substances, including caffeine.
This enzyme is a member of a large family of enzymes called cytochrome P450.
Presence of variants: likely higher coffee consumption.

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
 library(myDNA)
 library(plyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
  cofeeConsumption(myDNA)
}
}
\author{
Inga Patarcic
}
