% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{deepSleeper}
\alias{deepSleeper}
\title{Function to screen for deep sleep}
\usage{
deepSleeper(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for deep sleep
}
\details{
SOURCE: 23andMe report
INFO from 23andMe:
The genetic marker in this report is in the ADA gene,
which contains instructions for an enzyme that helps control adenosine levels.
Scientists think that adenosine builds up more quickly in people with one or
two copies of the T variant at this marker. This extra adenosine increases
sleepiness, leading to stronger delta waves. Because of this stronger sleep
pressure, people with the T variant also report feeling sleepier than other
people after a night of missed sleep.


NOTE! MyHeritage genotypes do not report this SNP, however missing genotypes
can be imputed and accessed by DNALand imputation tool

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
 library(myDNA)
 library(dplyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
  deepSleeper(myDNA)
}
}
\author{
Inga Patarcic
}
