% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_and_LiftoverF.R
\name{expandByLD}
\alias{expandByLD}
\title{For genotypes identify proxy SNPs (with R2 above threshold)}
\usage{
expandByLD(myDNA, pop = "CEU", R.squared = 0.8, chunks = 1, out.dir)
}
\arguments{
\item{myDNA}{GRanges object, output of importDNA function}

\item{pop}{(character, default EUR) the name of a 1000 Genomes population
(AMR,AFR,ASN,EUR,...). Set this to NA to use all populations.}

\item{R.squared}{(numeric) threshold for R.squared statistics used to
assess LD}

\item{chunks}{(numeric, default=1). Enables running function for N chunks
of SNPs. It is to intensive to run it on the full set of genotypes (700k)
so it is better to split it into cca 1000 chunks}

\item{out.dir}{(character) path to the directory where data will be stored}
}
\description{
For genotypes identify proxy SNPs (with R2 above threshold)
}
\details{
For analyzed genotypes (myDNA) this function identifies all SNPs
that in LD above predefined threshold. Since number of SNPs is large (>700k)
analysis is separated into chunks, and results are pooled together in the end
In addition, chunks and final results are saved in the out.dir
}
\examples{
\dontrun{}
}
\author{
Inga Patarcic
}
