% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeScreens.R
\name{identifyRiskAllele}
\alias{identifyRiskAllele}
\title{f which adds info if i have a risk or not}
\usage{
identifyRiskAllele(myDNAAdded, riskSNP = "STRONGEST.SNP.RISK.ALLELE")
}
\arguments{
\item{myDNAAdded}{data.table with genotypes and added info from snp-diease DB}

\item{riskSNP}{column name where risk SNP info is stores}
}
\description{
f which adds info if i have a risk or not
}
\author{
Inga Patarcic
}
\keyword{internal}
