% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importF.R
\name{importDNA}
\alias{importDNA}
\title{Function to import myHeritage, 23andMe, AncestryDNA, Illumina personal genotypes}
\usage{
importDNA(myGenotypes, type = "myHeritage", asGRanges = FALSE)
}
\arguments{
\item{myGenotypes}{(character) path to the genotype file}

\item{type}{(character) info about the company used to genotype data: OPTIONS
"myHeritage", "23andMe", "AncestryDNA", "Illumina"}

\item{asGRanges}{(default FALSE) Result of the function is GRanges or
data.frame object}
}
\description{
Function to import myHeritage, 23andMe, AncestryDNA, Illumina personal genotypes
}
\details{
Import genotypes into R
}
\examples{

\dontrun{
# example myHeritage
library(GenomicRanges)
library(stringr)
myGenotypes="D:/GoogleDrive/Projects/myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
myDNA <- importDNA(myGenotypes)
}
}
\author{
Inga Patarcic
}
