% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{lactoseIntolerance}
\alias{lactoseIntolerance}
\title{Function to screen for lactose intolerance}
\usage{
lactoseIntolerance(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for lactose intolerance
}
\details{
SOURCE: 23andMe report
INFO:
Lactose intolerance is influenced by a genetic marker near the LCT gene.
GG - likely lactose intolerant. CG/CC - likely not lactose intolerant

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
}
}
\author{
Inga Patarcic
}
