% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{muscleComposition}
\alias{muscleComposition}
\title{Function to screen for production of ACTN3 gene}
\usage{
muscleComposition(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for production of ACTN3 gene
}
\details{
SOURCE: 23andMe report
INFO:
A genetic marker in the ACTN3 gene.
This marker controls whether muscle cells produce a protein
(called alpha-actinin-3) that's found in fast-twitch muscle fibers.
While some people don't produce this protein at all, almost all of the elite
power athletes who have been studied have a genetic variant that allows them to
produce the protein. This suggests that the protein may be beneficial at least
at the highest levels of power-based athletic competition.

NOTE! MyHeritage genotypes do not report this SNP, however missing genotypes
can be imputed and accessed by DNALand imputation tool

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
 library(myDNA)
 library(plyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
 muscleComposition(myDNA)
}
}
\author{
Inga Patarcic
}
