% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeScreens.R
\name{myDNAScreenSNPS}
\alias{myDNAScreenSNPS}
\title{Overlaps myDNA genotypes with SNPs of interest}
\usage{
myDNAScreenSNPS(myDNA, snpsData)
}
\arguments{
\item{myDNA}{(data.frame) analyzed genotypes, an output of importDNA function}

\item{snpsData}{(data.frame) with min 2 columns: rsID (SNPid) and risk allele
In that order!}
}
\description{
Overlaps myDNA genotypes with SNPs of interest
}
\details{
This function identifies overlap between my genotypes (myDNA object)
and data.frame which stores info SNPs (their unique ID -rsID, and risk allele
-in that order. This function is useful when user already has a list of SNPs
associated with a trait of interest.
}
\examples{
# example - creating example for myDNAScreenSNPS()
# IDEA is as follows: extract rs and snp risk alleles, and extract those for
 coronary heart disease
 # filter CHD SNPS
\dontrun{dd.heart <- filter(db,
                   MAPPED_TRAIT=="coronary heart disease")

snpsData <-   dd.heart$STRONGEST.SNP.RISK.ALLELE
# greping risk allele
riskAlleles <- gsub("[^\\\\-]*-([ATCG?])", "\\\\1",
                    snpsData)

snpsData <- cbind(dd.heart$SNPS,riskAlleles)


# check my genome for given list of SNPs
myDNAScreenSNPS(myDNA, snpsData )

 }
}
