% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{neanderthalMe}
\alias{neanderthalMe}
\title{Function to screen for certain Neanderthal variants}
\usage{
neanderthalMe(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for certain Neanderthal variants
}
\details{
SOURCE 1: 23andMe report
If you have certain Neanderthal variants, it means that some of your physical
traits may trace back to your Neanderthal ancestors.

Tested variants/gene/trait:
rs7544462/Gene: MEAF6/Trait: Height
rs1877547/Gene: LPP/Trait: Height
rs4849721/Gene: Near the EN1 gene/Trait: Less back hair
rs12458349/Gene: Near the PHLPP1 gene/Trait: Straight hair
rs11213819/Gene: Near the C11orf53 gene/Trait: Less likely to
sneeze after eating dark chocolate

NOTE! MyHeritage genotypes do not report some if the SNPs, however missing
genotypes can be imputed and accessed by DNALand imputation tool

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
 library(myDNA)
 library(dplyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
 neanderthalMe(myDNA)
}
}
\author{
Inga Patarcic
}
