% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTests.R
\name{redHair}
\alias{redHair}
\title{Function to screen for  red hairiness}
\usage{
redHair(myDNA)
}
\arguments{
\item{myDNA}{(character) path to the genotype file}
}
\description{
Function to screen for  red hairiness
}
\details{
SOURCE 1: 23andMe report
NOTE! MyHeritage genotypes do not report some if the SNPs, however missing
genotypes can be imputed and accessed by DNALand imputation tool
INFO:
Several variants in a single gene, MC1R, can cause red hair by increasing
the amount of pheomelanin in your hair.
Hair color is determined by not just how much pigment you have,
but also what kind. Whether you have red hair depends on your levels of
red/yellow pigment (pheomelanin). But the lightness or darkness of your hair
depends on your levels of a brown/black pigment called eumelanin
REPORTED SNPS: rs1805007,rs1805008,i3002507

SOURCE 2: https://www.snpedia.com/index.php/I3002507
i3002507 alias 	rs1805009 identified in SNPedia
proxy calculations needed

!!!! myAllele column reports whether risk allele was identified in my genome
}
\examples{

\dontrun{
# example myHeritage
 library(myDNA)
 library(plyr)
 library(stringr)
  Genome="/data/akalin/Projects/AAkalin_myDNA/Data/MyHeritage/MyHeritage_raw_dna_dataInga/MyHeritage_raw_dna_data.csv"
   myDNA <- importDNA(myGenotypes = Genome,type = "myHeritage" )
 redHair(myDNA)
}
}
\author{
Inga Patarcic
}
