% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeScreens.R
\name{traitScreen}
\alias{traitScreen}
\title{Filter prescreened myDNA object filter for trait of interest}
\usage{
traitScreen(myScreenDNA, trait = "all", risk = "high.protective")
}
\arguments{
\item{trait}{(character) which trait are you interested in}

\item{myDNA}{GRanges object, output of importDNA function}
}
\description{
Filter prescreened myDNA object filter for trait of interest
}
\details{
This function allows you to filter your imported myDNA object +
overlapped with database of interest (GWASCatalog) by selecting trait in
which you are interested in, and selecting category of SNPs. For example,
high risk SNPs are those which have reported OR or beta >1.5.
}
\examples{
#' \dontrun{
myAchille <- traitScreen(myScreenDNA, risk="high")
myDepression <- traitScreen(myScreenDNA, risk="all",
trait="unipolar depression")
myDepression.high <- traitScreen(myScreenDNA, risk="high",
trait="unipolar depression")
}
}
\author{
Inga Patarcic
}
